+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Tests/RemoteDependencyDataNewTests.cs
@@ -121,5 +121,38 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Tests
 
             Assert.Equal(activity.DisplayName, remoteDependencyDataName);
         }
+
+        [Fact]
+        public void ValidateMessagingRemoteDependencyData()
+        {
+            using var tracerProvider = Sdk.CreateTracerProviderBuilder().AddSource(nameof(ValidateMessagingRemoteDependencyData)).Build();
+            using var activitySource = new ActivitySource(nameof(ValidateMessagingRemoteDependencyData));
+            using var activity = activitySource.StartActivity(
+                ActivityName,
+                ActivityKind.Producer,
+                parentContext: new ActivityContext(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded),
+                startTime: DateTime.UtcNow);
+            Assert.NotNull(activity);
+            activity.Stop();
+
+            activity.SetStatus(Status.Ok);
+            activity.SetTag(SemanticConventions.AttributeMessagingSystem, "servicebus");
+            activity.SetTag(SemanticConventions.AttributeServerAddress, "my.servicebus.windows.net");
+            activity.SetTag(SemanticConventions.AttributeMessagingDestinationName, "queueName");
+
+            var activityTagsProcessor = TraceHelper.EnumerateActivityTags(activity);
+
+            var remoteDependencyData = new RemoteDependencyData(2, activity, ref activityTagsProcessor);
+
+            Assert.Equal("RemoteDependencyDataNewActivity", remoteDependencyData.Name);
+            Assert.Equal(activity.Context.SpanId.ToHexString(), remoteDependencyData.Id);
+            Assert.Equal("my.servicebus.windows.net/queueName", remoteDependencyData.Data);
+            Assert.Null(remoteDependencyData.ResultCode);
+            Assert.Equal(activity.Duration.ToString("c", CultureInfo.InvariantCulture), remoteDependencyData.Duration);
+            Assert.Equal("my.servicebus.windows.net/queueName", remoteDependencyData.Target);
+            Assert.Equal(activity.GetStatus() != Status.Error, remoteDependencyData.Success);
+            Assert.True(remoteDependencyData.Properties.Count == 0);
+            Assert.True(remoteDependencyData.Measurements.Count == 0);
+        }
     }
 }