+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Tests/RequestDataNewTests.cs
@@ -39,19 +39,22 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Tests
             activity.SetTag(SemanticConventions.AttributeServerAddress, "www.foo.bar");
             activity.SetTag(SemanticConventions.AttributeUrlPath, "/search");
             activity.SetTag(SemanticConventions.AttributeHttpResponseStatusCode, null);
+            activity.SetTag("foo", "bar");
 
             var activityTagsProcessor = TraceHelper.EnumerateActivityTags(activity);
 
             var requestData = new RequestData(2, activity, ref activityTagsProcessor);
 
-            Assert.Equal("GET /search", requestData.Name);
+            // Name is set later via operation name on TelemetryItem
+            Assert.Null(requestData.Name);
             Assert.Equal(activity.Context.SpanId.ToHexString(), requestData.Id);
             Assert.Equal(httpUrl, requestData.Url);
             Assert.Equal("0", requestData.ResponseCode);
             Assert.Equal(activity.Duration.ToString("c", CultureInfo.InvariantCulture), requestData.Duration);
             Assert.False(requestData.Success);
             Assert.Null(requestData.Source);
-            Assert.True(requestData.Properties.Count == 0);
+            Assert.True(requestData.Properties.Count == 1);
+            Assert.Equal("bar", requestData.Properties["foo"]);
             Assert.True(requestData.Measurements.Count == 0);
         }
 
@@ -82,11 +85,15 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Tests
         }
 
         [Theory]
-        [InlineData("200", true)]
-        [InlineData("400", false)]
-        [InlineData("500", false)]
-        [InlineData("0", false)]
-        public void ValidateHttpRequestSuccess(string httpStatusCode, bool isSuccess)
+        [InlineData("200", OperationType.Http, true)]
+        [InlineData("400", OperationType.Http, false)]
+        [InlineData("500", OperationType.Http, false)]
+        [InlineData("0", OperationType.Http, false)]
+        [InlineData(null, OperationType.Http, true)]
+        [InlineData("", OperationType.Http, true)]
+        [InlineData("someStatusCode", OperationType.Unknown, false)] // Activity status is set to error in the test code for validation.
+        [InlineData("someStatusCode", OperationType.Messaging, true)]
+        internal void ValidateHttpRequestSuccess(string httpStatusCode, OperationType operationType, bool isSuccess)
         {
             using var tracerProvider = Sdk.CreateTracerProviderBuilder().AddSource(nameof(ValidateHttpRequestSuccess)).Build();
             using var activitySource = new ActivitySource(nameof(ValidateHttpRequestSuccess));
@@ -98,16 +105,12 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Tests
             Assert.NotNull(activity);
             activity.Stop();
 
-            var httpResponseCode = httpStatusCode ?? "0";
-            activity.SetTag(SemanticConventions.AttributeHttpRequestMethod, "GET");
-            activity.SetTag(SemanticConventions.AttributeHttpResponseStatusCode, httpStatusCode);
+            if (operationType == OperationType.Unknown)
+            {
+                activity.SetStatus(ActivityStatusCode.Error);
+            }
 
-            var activityTagsProcessor = TraceHelper.EnumerateActivityTags(activity);
-
-            var requestData = new RequestData(2, activity, ref activityTagsProcessor);
-
-            Assert.Equal(httpResponseCode, requestData.ResponseCode);
-            Assert.Equal(isSuccess, requestData.Success);
+            Assert.Equal(isSuccess, RequestData.IsSuccess(activity, httpStatusCode, operationType));
         }
 
         [Fact]
@@ -218,6 +221,42 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Tests
             var startTimeEpoch = startTime.ToUnixTimeMilliseconds();
 
             Assert.False(requestData.Measurements.TryGetValue("timeSinceEnqueued", out var timeInQueue));
+        }
+
+        [Fact]
+        public void ValidateMessagingRequestData()
+        {
+            using var tracerProvider = Sdk.CreateTracerProviderBuilder().AddSource(nameof(ValidateMessagingRequestData)).Build();
+            using var activitySource = new ActivitySource(nameof(ValidateMessagingRequestData));
+            using var activity = activitySource.StartActivity(
+                ActivityName,
+                ActivityKind.Consumer,
+                parentContext: new ActivityContext(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded),
+                startTime: DateTime.UtcNow);
+            Assert.NotNull(activity);
+            activity.Stop();
+
+            activity.SetStatus(Status.Ok);
+            activity.SetTag(SemanticConventions.AttributeMessagingSystem, "servicebus");
+            activity.SetTag(SemanticConventions.AttributeServerAddress, "my.servicebus.windows.net");
+            activity.SetTag(SemanticConventions.AttributeMessagingDestinationName, "queueName");
+            activity.SetTag("foo", "bar");
+
+            var activityTagsProcessor = TraceHelper.EnumerateActivityTags(activity);
+
+            var requestData = new RequestData(2, activity, ref activityTagsProcessor);
+
+            // Name is set later via operation name on TelemetryItem
+            Assert.Null(requestData.Name);
+            Assert.Equal(activity.Context.SpanId.ToHexString(), requestData.Id);
+            Assert.Equal("my.servicebus.windows.net/queueName", requestData.Url);
+            Assert.Equal("0", requestData.ResponseCode);
+            Assert.Equal(activity.Duration.ToString("c", CultureInfo.InvariantCulture), requestData.Duration);
+            Assert.True(requestData.Success);
+            Assert.Equal("my.servicebus.windows.net/queueName", requestData.Source);
+            Assert.True(requestData.Properties.Count == 1);
+            Assert.Equal("bar", requestData.Properties["foo"]);
+            Assert.True(requestData.Measurements.Count == 0);
         }
 
         private ActivityLink AddActivityLink(long enqueuedTime)