+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Tests/RequestDataTests.cs
@@ -50,20 +50,23 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Tests
             activity.SetTag(SemanticConventions.AttributeHttpMethod, "GET");
             activity.SetTag(SemanticConventions.AttributeHttpRoute, "/search");
             activity.SetTag(SemanticConventions.AttributeHttpUrl, httpUrl); // only adding test via http.url. all possible combinations are covered in AzMonListExtensionsTests.
-            activity.SetTag(SemanticConventions.AttributeHttpStatusCode, null);
+            activity.SetTag(SemanticConventions.AttributeHttpStatusCode, "200");
+            activity.SetTag("foo", "bar");
 
             var activityTagsProcessor = TraceHelper.EnumerateActivityTags(activity);
 
             var requestData = new RequestData(2, activity, ref activityTagsProcessor);
 
-            Assert.Equal("GET /search", requestData.Name);
+            // Name is set later via operation name on TelemetryItem
+            Assert.Null(requestData.Name);
             Assert.Equal(activity.Context.SpanId.ToHexString(), requestData.Id);
             Assert.Equal(httpUrl, requestData.Url);
-            Assert.Equal("0", requestData.ResponseCode);
+            Assert.Equal("200", requestData.ResponseCode);
             Assert.Equal(activity.Duration.ToString("c", CultureInfo.InvariantCulture), requestData.Duration);
-            Assert.False(requestData.Success);
+            Assert.True(requestData.Success);
             Assert.Null(requestData.Source);
-            Assert.True(requestData.Properties.Count == 0);
+            Assert.True(requestData.Properties.Count == 1);
+            Assert.Equal("bar", requestData.Properties["foo"]);
             Assert.True(requestData.Measurements.Count == 0);
         }