+++ sdk/resourcemanager/Azure.ResourceManager/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.7.0-beta.1 (Unreleased)
+## 1.8.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,13 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.7.0 (2023-07-13)
+
+### Other Changes
+
+- Bump api-version of `Subscription` to `2022-12-01`
+- Bump api-version of `Tenant` to `2022-12-01`
 
 ## 1.6.0 (2023-05-16)